feat(sheets): form submissions on rows; fix second review round
## Form submissions

`appendFormSubmission` wrote a row by reparsing the whole sheet, splicing the
row in and re-serialising all of it — so a submission cost O(sheet), seconds of
CPU once a form had collected tens of thousands of responses, which is exactly
when a form is working. It is one addressed cell write now.

Two things fell out of that. The unreadable-sheet guards are gone with the
read-modify-write they protected: there is no document parse to fail, and no
path by which one submission can touch another row. And the retry loop is gone
because it became unreachable — it existed for the `expectedRevision` race
between two forms targeting one sheet, and addressed writes to different rows do
not contend, while the `FOR UPDATE` lock on `form_targets` already serialises
submissions to the same form. Removed rather than left as decoration.

The submitter IP hash still had to be recorded — it is the audit trail for an
anonymous, publicly reachable write. It now rides a `logActivityWithTx` entry
inside the same transaction, carrying the metadata without the document.

## Second review round

- A range formula wider than ~65k rows made every edit to its inputs fail.
  `evaluateClosure` expanded each dependency rectangle into one row index per
  row and handed the lot to `IN (...)`, rebuilding the bind-parameter ceiling in
  the read path — on precisely the 100k-row sheet this design exists to support,
  and after the same ceiling had already been fixed on the write path. Ranges
  now reach SQL as bounds: two parameters however wide. `loadRowsByIndex`
  chunks, and `rebuildTab`'s delete no longer builds an unbounded ARRAY.

- `delete-rows` never rebuilt. Shifting indexes moves cells without changing
  their text, so every dependent was stale and every dependency edge named a
  moved or deleted address — and `rebuildTab`, the documented remedy, had no
  production caller. `deleteRows` now rebuilds itself.

- `applyEvaluation` could invent a cell that was never authored, given a stale
  edge pointing at a deleted address.

- `deleteRows` shrank the extent and reported `deleted: count` even when it
  matched nothing, so `delete-rows {fromRow: 5000, count: 100}` on a 100-row
  sheet reported success and rendered the sheet empty.

- `queryRows` paginated non-deterministically whenever the sort key had ties,
  so an agent reading all matching rows could see one twice and miss another.
  `rowIndex` now always tie-breaks.

- `neq` silently dropped rows whose cell was absent: `NULL <> 'active'` is NULL,
  so "status is not active" excluded every blank row.

- `sheet_snapshots.seq` was `integer` against a shared `bigserial` sequence.

- The integration suite wiped `users`/`drives`/`pages` wholesale, which would
  empty the shared CI database out from under the 26 other suites now in the
  same step. Cleanup is row-scoped to what each test seeded.

- `--page` with a missing or flag-shaped argument silently backfilled every
  sheet in the database instead of one.

+++ packages/db/drizzle/0269_fair_expediter.sql
@@ -48,7 +48,7 @@ CREATE TABLE "sheet_rows" (
 CREATE TABLE "sheet_snapshots" (
 	"id" text PRIMARY KEY NOT NULL,
 	"pageId" text NOT NULL,
-	"seq" integer NOT NULL,
+	"seq" bigint NOT NULL,
 	"contentRef" text NOT NULL,
 	"contentFormat" "content_format",
 	"contentSize" integer,
